frontend/requirements.txt
cloudinary==1.44.0
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.1.0
requests==2.32.3
Werkzeug==3.1.3